ListTag: Use a linked-list for value storing, add several new API methods
This now internally uses an SplDoublyLinkedList, which has a range of advantages over a typical array.
The primary advantage being that keys are always contiguous and are not tied to specific values like they are in an array.
It makes it far more convenient to insert/remove tags midway through a list, push/pop them and unshift/shift.

This adds a range of new API methods for ListTags: push(), pop(), shift(), unshift(), insert(), remove(), get(), set(), first(), last(), isset() and empty().
These should make it much more comfortable to work with ListTags and is really nice after dealing with array-access for so long (which didn't even support appending tags!!!)

ListTags now also implement the \Iterator interface and won't throw shit out at you like the __name property when iterating on them.

The `$mode` parameter of count() has been removed, since \Countable does not declare or receive this parameter.