ListTag: enforce type-checking for adding members
This is a much-needed change and will make ListTags much less error-prone.

When adding new NamedTags to a ListTag in any way, its type will now be checked. If the list's type is TAG_End and the list is empty, inserting a new tag will cause the type to be automatically updated.
Otherwise, a TypeError will be thrown when adding a tag of the wrong type to a list.

ListTags' member type also can no longer be modified when the list is not empty, for obvious reasons.