ListTag: fix compatibility with very old NBT Some old NBT implementations used to signify an empty list using a type of TAG_Byte, which is very confusing. This would cause bugs in this implementation because empty lists with that type (intended to indicate empty list) would only be able to have TAG_Bytes in them, which is undesired behaviour.
This commit works around issues like this by automatically setting the list tag type to TAG_End if the list is empty when decoded. Since TAG_End is not a valid tag type and indicates that the type should be auto-detected, this resolves possible issues with old data.

HOWEVER, THERE IS ONE CAVEAT:
If used to decode and subsequently encode old data, this will NOT produce the same data as was originally put in if any TAG_Lists had non-TAG_End types on empty lists. However, this is consistent with how MC Java handles this (they do the same, set tag type to 0 on an empty list).