Revert "CompoundTag::getByte() and friends now return null for missing tags"
This reverts commit 16afd20c6f9e1e32ef76ea2f282f2493c32987db.

In practice, this is a major pain in the ass, because the majority case
is the assumption that a tag will always be present. In such cases, you
usually want to bail out anyway if there's something wrong, so the
original behaviour makes more sense.

It's possible we could make two sets of methods, e.g. getByte() vs
mustGetByte(), one which must be given a default, and the other which
doesn't. However, we got this far without such a mess, so there's not
really much of a reason to start now.