ListTag: use array instead of SplDoublyLinkedList the idea of using linked-list for this was to allow dynamic insertions and deletions midway through the list.
however, this is basically never done in practice, and because SplDoublyLinkedList has O(n) read performance,
we pay a read performance penalty for these write features which are basically never used anyway.

in the future, we should perhaps consider getting rid of the mutation APIs from ListTag, and require creating
a new one to replace it in mutation cases, similar to other tags (aside from TAG_Compound).